+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## unreleased
 
+- fix: Drop Sessions without release name #826
 - feat: Bring back SentryOptions.enabled #818
 - fix: Remove enum specifier for SentryLevel #822
 - feat: Send environment 'production' if nothing was set #825